+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 
 
+### Add 2 new HA Features
+
+> Branches Affected: HA  
+> Tags Affected:   
+> Date: Fri, 8 Mar 2019 10:14:31 -0500  
+> Author: Tyler Moore (tmoore@goflyball.com)  
+> Committer: Tyler Moore (tmoore@goflyball.com)  
+
+- mysql galera replication
+- pacemaker / corosync cluster with floating ip
+
+
+---
+
+
 ### Add Mysql Replication Scripts
 
 > Branches Affected: HA